Log level matching refactor (#45)
`Captured` now takes an explicit `levelname` and `levelno`. This allows
3rd-party logging framework integrations to be freed of the obligation
to match stdlib `logging` levels.

The validation of `validate_levelno` was relaxed to just validate the
type of the input, and correspondingly renamed to `validate_level`.

The `DEFAULT_XXX` class attributes on `Logot` were updated to prefer
`str` and use string constants for easier reading in docs. The typing of
log levels everywhere was updated to prefer `str` for better
compatibility with 3rd-party logging frameworks.

The named logging helpers in `logged` were updated to use string level
names for better compatibility with 3rd-party logging frameworks.

Needed by #28.